refactor(whispering): rename (app)/_layout-utils to _runtime
The folder name framed these as helpers for the layout component, but they
are the app's runtime wiring (update check, onboarding, accessibility,
shortcut registration, recorder-window sync). Rename to say what they are.
Pure rename plus import-path updates; no behavior change.
