+++ apps/whispering/src/routes/(app)/_runtime/check-for-updates.ts
