+++ apps/whispering/src/routes/(app)/_runtime/register-accessibility-permission.ts
