+++ apps/whispering/src/routes/(app)/_runtime/register-commands.ts
