+++ apps/whispering/src/routes/(app)/_runtime/syncIconWithRecorderState.svelte.ts
